Cleaned up file structure, created readmes'
I cleaned up the file structure for repository so that older files are now placed in the misc folder. Then I edited the readme for the month of August. My goal is the create a readme for each day and month, so that I may keep a better record of the progress I am needing to make.