Wrote some for yesterdays challenge
Though I wasn't able to complete the code for this challenge in a day, I was testing different solutions from a variety of languages including C, Java, and Javascript. I can't seem to understand how to map out and give them value. I believe I need to get more practice and these challenges are to complex for me to solve in a day.